added Section class

+++ src/main/java/com/iss/cms/domain/Conference.java
@@ -3,7 +3,6 @@ package com.iss.cms.domain;
 import lombok.*;
 import javax.persistence.*;
 import java.time.LocalDate;
-import java.util.ArrayList;
 
 @Entity
 @NoArgsConstructor
@@ -12,10 +11,9 @@ import java.util.ArrayList;
 public class Conference {
     @Id
     @GeneratedValue
-    private Long id;
+    private int id;
 
     private String name;
     private LocalDate date;
     private int entryFee;
-//    private ArrayList<String> sections;
 }
